pyproject.toml
[tool.poetry]
include = [
    { path = "pumaguard-sounds/cougar_call.mp3", format = [ "sdist", "wheel" ] },
    { path = "pumaguard-sounds/deterrent_puma.mp3", format = [ "sdist", "wheel" ] },
]
name = "pumaguard"
version = "0.0.0"
description = ""
authors = ["Nicolas Bock <nicolasbock@gmail.com>"]
readme = "README.md"
homepage = "http://pumaguard.rtfd.io/"
repository = "https://github.com/PEEC-Nature-Youth-Group/pumaguard"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
packaging = "~=25.0"
pandas = "~=2.3"
pillow = "~=12.0"
pyyaml = "~=6.0"
tensorflow = "~=2.20"
ultralytics = "~=8.3"
requests = "~=2.32"

[tool.poetry.group.dev.dependencies]
ansible = "~=10.7"
ansible-lint = { version = "~=25.2", markers = 'platform_system != "Windows"' }
bashate = "~=2.1"
black = "~=25.9"
ipykernel = "~=6.29"
isort = "~=6.0"
mypy = "~=1.18"
opencv-python = "~=4.11"
passlib = "~=1.7"
plotly = "~=6.0"
pylint = "~=3.3"
pytest = "~=8.3"
pytest-cov = "~=7.0"
scipy = "~=1.15"

[tool.poetry.extras]
and-cuda = ["tensorflow-and-cuda"]

[tool.poetry.scripts]
pick-files = 'scripts.pick_files:main'
pumaguard = 'pumaguard.main:main'
playsound = 'pumaguard.sound:main'
pumaguard-models = 'pumaguard.model_cli:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = true
include_trailing_comma = true
multi_line_output = "VERTICAL_HANGING_INDENT"
py_version = 310
split_on_trailing_comma = true

[tool.black]
line-length = 79

[tool.ruff]
line-length = 79

[tool.basedpyright]
reportUnusedCallResult = "hint"
